requirements.txt
autogen-agentchat>=0.5.7,<0.8.0
autogen-ext[openai]>=0.5.7,<0.8.0
fastapi>=0.115,<1.0
httpx>=0.27,<1.0
jinja2>=3.1,<4.0
python-dotenv>=1.0,<2.0
python-multipart>=0.0.9,<1.0
requests>=2.32,<3.0
uvicorn[standard]>=0.30,<1.0
pytest>=8.0,<9.0
